handling breakpoint related actions
